Use anys in the result/maybe types
This is because when using unknown, a function which has multiple return
statements such that some of them are return Result.err and some
Result.ok, will confuse TS.

Ultimately TS won't be able to figure out that the two are related and
will simply assume that the function returns a union of types rather
than a single unified type because unknown does not allow for the
specific other type while any seems to. Or at least that's my
understanding. Might not be 100% accurate but it works